pyproject.toml
[build-system]
requires = ["setuptools ~= 65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fenicsx_pctools"
version = "0.5.0.dev0"
description = "FEniCSx Preconditioning Tools"
authors = [
    {name = "Martin Rehor", email="martin.rehor@rafinex.com"},
    {name = "Jack S. Hale", email = "jack.hale@uni.lu"}
]
license = {file = "COPYING"}
readme = "README.rst"
requires-python = ">=3.7"

keywords = ["FEniCS", "preconditioning", "PETSc", "block systems"]

classifiers = [
    "Topic :: Software Development"
]

dependencies = [
    "fenics-dolfinx>=0.5<0.6",
    "petsc4py"
]

[project.optional-dependencies]
docs = [
    "jupytext",
    "sphinx",
    "myst-parser"
]
examples = [
    "matplotlib",
    "pandas",
    "dolfiny @ git+https://github.com/michalhabera/dolfiny.git#egg=dolfiny"
]
lint = [
    "black",
    "flake8",
    "isort"
]
ci = [
    "fenicsx_pctools[docs]",
    "fenicsx_pctools[examples]",
    "fenicsx_pctools[lint]",
    "pytest"
]

[tool.setuptools]
packages = ["fenicsx_pctools"]

[project.urls]
repository = "https://gitlab.com/rafinex-external-rifle/fenicsx_pctools"

[tool.isort]
src_paths = ["demo", "examples", "fenicsx_pctools"]
known_first_party = ["basix", "dolfinx", "fenicsx_pctools", "ffcx", "ufl"]
known_third_party = ["gmsh", "numpy", "pytest"]
known_mpi = ["mpi4py", "petsc4py"]
sections= ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "MPI", "LOCALFOLDER"]
